Started work on refactoring LCD display class
 - Preliminary work, not nearly complete. Need to implement tests.
