+++ requirements.txt
@@ -1,6 +1,7 @@
 certifi==2019.3.9
 chardet==3.0.4
 idna==2.8
+Pillow==6.0.0
 python-decouple==3.1
 requests==2.22.0
 typing==3.6.6
